+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetitemop.py
@@ -1440,7 +1440,7 @@ class Evaluations3(BaseModel):
 
     source: Optional[CreateDatasetItemEvaluationsDatasetsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-06-15T05:10:28.298Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-06-15T08:56:18.245Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1508,7 +1508,7 @@ class Evaluations2(BaseModel):
 
     source: Optional[CreateDatasetItemEvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-06-15T05:10:28.298Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-06-15T08:56:18.244Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1576,7 +1576,7 @@ class Evaluations1(BaseModel):
 
     source: Optional[EvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-06-15T05:10:28.297Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-06-15T08:56:18.243Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1671,7 +1671,7 @@ class CreateDatasetItemResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-06-15T05:10:09.577Z")
+    updated: Optional[datetime] = parse_datetime("2026-06-15T08:55:55.775Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")