+++ packages/orq-rc/src/orq_ai_sdk/models/getallmemorystoresop.py
@@ -22,6 +22,10 @@ class GetAllMemoryStoresRequestTypedDict(TypedDict):
     r"""A cursor for use in pagination. `starting_after` is an object ID that defines your place in the list. For instance, if you make a list request and receive 20 objects, ending with `01JJ1HDHN79XAS7A01WB3HYSDB`, your subsequent call can include `after=01JJ1HDHN79XAS7A01WB3HYSDB` in order to fetch the next page of the list."""
     ending_before: NotRequired[str]
     r"""A cursor for use in pagination. `ending_before` is an object ID that defines your place in the list. For instance, if you make a list request and receive 20 objects, starting with `01JJ1HDHN79XAS7A01WB3HYSDB`, your subsequent call can include `before=01JJ1HDHN79XAS7A01WB3HYSDB` in order to fetch the previous page of the list."""
+    search: NotRequired[str]
+    r"""Filter memory stores by key (case-insensitive match)"""
+    updated_by: NotRequired[str]
+    r"""Filter by the users who last updated the memory store. Accepts a comma-separated list of user IDs"""
 
 
 class GetAllMemoryStoresRequest(BaseModel):
@@ -43,9 +47,23 @@ class GetAllMemoryStoresRequest(BaseModel):
     ] = None
     r"""A cursor for use in pagination. `ending_before` is an object ID that defines your place in the list. For instance, if you make a list request and receive 20 objects, starting with `01JJ1HDHN79XAS7A01WB3HYSDB`, your subsequent call can include `before=01JJ1HDHN79XAS7A01WB3HYSDB` in order to fetch the previous page of the list."""
 
+    search: Annotated[
+        Optional[str],
+        FieldMetadata(query=QueryParamMetadata(style="form", explode=True)),
+    ] = None
+    r"""Filter memory stores by key (case-insensitive match)"""
+
+    updated_by: Annotated[
+        Optional[str],
+        FieldMetadata(query=QueryParamMetadata(style="form", explode=True)),
+    ] = None
+    r"""Filter by the users who last updated the memory store. Accepts a comma-separated list of user IDs"""
+
     @model_serializer(mode="wrap")
     def serialize_model(self, handler):
-        optional_fields = set(["limit", "starting_after", "ending_before"])
+        optional_fields = set(
+            ["limit", "starting_after", "ending_before", "search", "updated_by"]
+        )
         serialized = handler(self)
         m = {}